Package states renamed into dialog_states
